Make the By implementations internal, and move the ByJQuery class into a different namespace.
