+++ src/Bumblebee.IntegrationTests/Specifications/ByJQueryTests/Given_I_am_searching_for_a_single_element.cs
@@ -2,9 +2,9 @@ using System;
 
 using Bumblebee.IntegrationTests.Shared.Hosting;
 using Bumblebee.IntegrationTests.Shared.Pages;
+using Bumblebee.JQuery;
 using Bumblebee.Setup;
 using Bumblebee.Setup.DriverEnvironments;
-using Bumblebee.Specifications;
 
 using FluentAssertions;
