+++ src/Bumblebee.IntegrationTests/Specifications/ByJQueryTests/Given_an_invalid_selector.cs
@@ -2,9 +2,9 @@ using System;
 
 using Bumblebee.IntegrationTests.Shared.Hosting;
 using Bumblebee.IntegrationTests.Shared.Pages;
+using Bumblebee.JQuery;
 using Bumblebee.Setup;
 using Bumblebee.Setup.DriverEnvironments;
-using Bumblebee.Specifications;
 
 using FluentAssertions;
